Switch from docopt to click. Sadly, after a few weeks of on and off attempts of getting docopt to work sanely with PyInstaller, I called it quits in favor of the (relatively new) click library from the fine folks at pocoo of Jinja2 and Babel fame.
Click is at http://click.pocoo.org/3/ and, like docopt, it
requires minimal code to use. Unlike docopt and the previous
implementation, it doesn't need to shell out another subprocess to
call a subcommand, and it also easily supports types and defaults –
and supports being frozen by PyInstaller out of the box.