Updating the way that translation spreadsheets are updated.
Previously all translation information was blindly written over the existing data.
This caused a couple of issues such as when several translations are added past the buffer cells
or the comments in a spreadsheet keeping the same position but having a different cell value
hence losing the context of the comment.

The updates are focused on being a bit more selective on how it updates the values already present,
adds new values, and removes obsolete messages. By being purposeful with the order of operations
the comments or other cell specific meta information is preserved.