Initial functionality for supporting the dynamic ui tools through config.
Reads the list of tools from the podspec and uses it to include the correct files from the project's `node_modules` directory.

This initial version is very opinionated and targeted towards the usage of npm. Future versions should make this path variable based on configuration in the podspec.

Part of #342