Adding the idea of a stream of imported docs.
Working on a way to create a depency graph from how the pages get built.

Each time a doc is imported into another template the stream will add the doc to the set.

After the render is complete the stream is popped.

The use of a global variable is a hack to get it to work for now. Ideally it would be created for each page that is rendered, but with the tags being added as part of the global env it makes it difficult to use a separate instance for each render.

This also would prevent the ability to thread the building since they would interfere with the global stream.