Switching the dependency stream to be a local variable.
Switched the `g.doc` and `g.docs` tags to be created during each render instead of using a global environment version.

Allows for creating the dependency stream for each render call and binding the tag usage to a local stream. Should be able to be used in a threaded rendering situation.