Reworking the way that the router is updated after file changes.
Instead of making the file_watchers do the updating of the pod there is a new `file_updated` method that determines what actions the pod needs to do when a file is updated.

Also changed how the routes are done when files are changing. Now the dependency graph is used to selectively remove routes from the serving rules and add the updated files back in. This allows for the `path` of the content file to be updated without recreating the entire routing map.

Part of #391