Adding a `locale_safe` property to the document.
This is a 'safe' way to reference what the locale of the document should be. It does not take into account any custom locale information defined in the front matter.

If it tries to use the normal `document.locale` while loading front-matter (such as `!g.doc`) it causes an infinite loop to use the `doc.locale.` There was a bug when just using the kwarg for the locale so this property uses the kwarg or falls back to the collection default.