Fixing a bug with the caching of documents with a default locale.
When caching it was trying to be smart and reuse the same loaded doc if there was no `locale` kwarg provided to load the doc. This had a side effect when a document contains a different locale than the collection default.

By checking when creating the routes we can determine if the doc should have been loaded with a different locale and redo the doc loading using the document specific default locale.