Reworking the routes cleanup to fix cache and redo the default locale.
There were issues with the document being cached in the `None` locale or the collection default when it wasn't the correct document. When making the routes we can examine to see when the default locale for the collection does not match the document default and can adjust the cache and add in the correct documents.