Making the option to use the new routing in `run`.
Can use side by side with the older rendering pipeline to be able to test.

Adding a render controller for managing the rendering process for different content types.

Adding a render pool for being able to thread the rendering later on.

Moved the RenderedDocument to be with the new `rendering` directory.