Fixing the route loading to better handle localized filenames.
When using localized filenames (ex: `content@ru.md` without a default filename (ex: `content.md`) the routes would ignore the localized pod path and not add it to the routes.

Documents that do not have a version with the default locale were also being ignored.

The docs loader also needs to verify that the localized documents exist before adding them to the routes.

Fixes #778