Adding the ability to export/import the routes information.
This allows for the routes to be generated and stored in a json file then for later commands to use the routes file to load routes instead of needing to regenerate all of the routing info.

Especially used for sharding processes as larger sites can take several minutes to load routing information.