Updating the yaml parse to separate into a base and normal loader.
Certain cases for the yaml loading should use a subset of the yaml constrcutors. Specifically ones that don't access collections and documents to prevent circular references.

This change allows for yaml to be loaded with some of the constructors (like `!g.yaml`) but not allow for using doc or static constructors which blow up due to circular references when loading thing such as collection blueprints.