Fix translation import logging
Ensure translation logging is consistent across translation import
methods. Previously, when using the local translations import command,
there was no output. A side effect of this change is that we no longer
have a progress bar for running imports, but that can be fixed later if
it is needed. Instead, we have real-time logging of translation import
differences